Worked on documentation, refactored enricher modules
There are now two groups of enricher which we deliver

* `fmp-` : Standard enrichers useful for creating default objects and annotations
* `f8-` : fabric8 specific enrichments used in the Fabric8 consiole and Fabric8 CD
