+++ core/src/main/java/io/fabric8/maven/core/util/MavenUtil.java
@@ -95,7 +95,7 @@ public class MavenUtil {
 
     public static String createDefaultResourceName(MavenProject project, String ... suffixes) {
         String suffix = StringUtils.join(suffixes, "-");
-        String ret = project.getArtifactId() + (suffix.length() > 0 ? "-" + suffix : suffix);
+        String ret = project.getArtifactId() + (suffix.length() > 0 ? "-" + suffix : "");
         if (ret.length() > 24) {
             ret = ret.substring(0,24);
         }
