--- enricher/git/src/main/resources/META-INF/fabric8/enricher-default
@@ -1,1 +1,0 @@
-io.fabric8.maven.enricher.git.GitEnricher
