--- enricher/misc/src/main/resources/META-INF/fabric8/enricher-default.txt
@@ -1,1 +1,0 @@
-io.fabric8.maven.enricher.misc.DeploymentAnnotationsToPodSpecEnricher
