+++ enricher/standard/src/main/resources/META-INF/fabric8/enricher-default
@@ -1,15 +1,15 @@
 # Default enrichers
 # =================
 
-# Add Maven coordinates as labels
-io.fabric8.maven.enricher.standard.ProjectEnricher
-
-# Add a default ReplicaSet (or ReplicationController if none is given)
+# Add a default Deployment, ReplicaSet or ReplicationController if none is given
 io.fabric8.maven.enricher.standard.ControllerEnricher
 
 # Add a default service if none is given. Enrich also with
 # other information found
 io.fabric8.maven.enricher.standard.ServiceEnricher
+
+# Add Maven coordinates as labels
+io.fabric8.maven.enricher.standard.ProjectEnricher
 
 # Add a default name for any resource missing
 io.fabric8.maven.enricher.standard.NameEnricher
@@ -19,3 +19,12 @@ io.fabric8.maven.enricher.standard.NameEnricher
 # in an given external descriptor (if more than one images are used within
 # a pod)
 io.fabric8.maven.enricher.standard.ImageEnricher
+
+# Copy over annotation from a deployment to its pod spec
+io.fabric8.maven.enricher.standard.PodAnnotationEnricher
+
+# Add SCM information found in .git as annotations
+io.fabric8.maven.enricher.standard.GitEnricher
+
+# Add all objects found in dependencies' jar within META-INF/fabric8/kubernetes.yml
+io.fabric8.maven.enricher.standard.DependencyEnricher
